backend/requirements.txt
# Core AI Framework
crewai>=0.80.0
crewai-tools>=0.12.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Document Processing
pypdf>=4.2.0
python-docx>=1.1.2
reportlab>=4.0.9

# PDF Generation
reportlab>=4.0.9

# Environment & Config
python-dotenv>=1.0.1

# Data Validation
pydantic>=2.8.2

# HTTP Client
requests>=2.31.0
httpx>=0.26.0

# Performance
aiofiles>=23.2.1

# Production Server
gunicorn>=21.2.0
